Remove dynamic metadata from global vars
The global vars used in this library are private and not used as
dynamic vars in the source, so they should not be marked as :dynamic.

They should also not have *earmuffs*, as that naming convention is
reserved for dynamic vars.

In order to prevent conflicts with local bindings, *qlocks* was
renamed to queue-locks.